icsToJson/requirements.txt
requests
icalendar
python-dateutil

